+++ tests/unit/TransactionTransferTest.php
@@ -5,10 +5,13 @@ namespace Test\Unit;
 require_once dirname(__DIR__) . '/Setup.php';
 
 use Braintree;
+use DateTime;
 use Test\Setup;
 
 class TransactionTransferTest extends Setup
 {
+    const TRANSFERTYPES = ['account_to_account', 'person_to_person', 'wallet_transfer', 'fund_transfer', 'fund_disbursement', 'payroll_disbursement', 'prepaid_top_up'];
+
     private function mockTransactionGatewayDoCreate()
     {
         return $this->getMockBuilder('Braintree\TransactionGateway')
@@ -21,21 +24,51 @@ class TransactionTransferTest extends Setup
     {
         $transactionGateway = $this->mockTransactionGatewayDoCreate();
         $transactionGateway
-            ->expects($this->once())
+            ->expects($this->exactly(7))
             ->method('_doCreate')
             ->will($this->returnCallback(function ($path, $params) {
-                $this->assertEquals("wallet_transfer", $params["transaction"]["transfer"]["type"]);
+                \PHPUnit\Framework\Assert::assertArrayHasKey('transfer', $params['transaction']);
+                \PHPUnit\Framework\Assert::assertArrayHasKey('type', $params['transaction']['transfer']);
+                \PHPUnit\Framework\Assert::assertContains(
+                    $params['transaction']['transfer']['type'],
+                    self::TRANSFERTYPES
+                );
             }));
 
-        $transactionParams = [
-        'type' => 'sale',
-        'amount' => '100.00',
-        'transfer' => [
-            'type' => "wallet_transfer",
-        ],
-        ];
-
-        $transactionGateway->sale($transactionParams);
+        foreach (self::TRANSFERTYPES as $type) {
+            $transactionParams = [
+                'type' => 'sale',
+                'amount' => '100.00',
+                'transfer' => [
+                    'type' => $type,
+                    'sender' => [
+                        'firstName' => 'Alice',
+                        'middleName' => 'A',
+                        'lastName' => 'Silva',
+                        'accountReferenceNumber' => '1000012345',
+                        'address' => [
+                            'streetAddress' => '1st Main Road',
+                            'locality' => 'Los Angeles',
+                            'region' => 'CA',
+                            'countryCodeAlpha2' => 'US',
+                        ],
+                        'dateOfBirth' => DateTime::createFromFormat('Y-m-d', '2012-04-10'),
+                    ],
+                    'receiver' => [
+                        'firstName' => 'Bob',
+                        'middleName' => 'A',
+                        'lastName' => 'Souza',
+                        'address' => [
+                            'streetAddress' => '2nd Main Road',
+                            'locality' => 'Los Angeles',
+                            'region' => 'CA',
+                            'countryCodeAlpha2' => 'US',
+                        ],
+                    ],
+                ],
+            ];
+            $transactionGateway->sale($transactionParams);
+        }
     }
 
     public function testSaleAcceptsTransferTypeOptions()